move: test cases to new folder
